Introduce libsodium tests
This commit introduces changes required for running libsodium's tests
on Unikraft. Tests can be enabled through the newly introduced kconfig
option CONFIG_LIBSODIUM_TESTS.

Minor changes on sodium's tests are introduced to allow building all
tests in a single Unikraft application. The core function that invokes
the test vectors in cmptest.h has been replaced in order to remove the
filesystem dependency for storing the test results.

The tests enabled are the ones executed when compiling libsodim natively
with --enable-minimal and running "make check". No test vectors have
been modified with the exception of sodium_core, sodium_util2, and
sodium_util3 to disable calls to exit() which cause the termination of
the Unikraft application.

Signed-off-by: Michalis Pappas <michalis.pappas@opensynergy.com>